\name{Transition-package}
\alias{Transition-package}
\alias{Transition}
\docType{package}
\title{\packageTitle{Transition}}
\description{\packageDescription{Transition}}
\section{Package Content}{\packageIndices{Transition}}
\author{\packageAuthor{Transition}}
\section{Maintainer}{\packageMaintainer{Transition}}
\keyword{package}
